+++ reference/html-proto/js/triggers.js
@@ -266,6 +266,9 @@ const _ARCHETYPE_BY_SIG = {
   'card_zone_change | card_is_creature, card_moves(battlefield, graveyard)': 'anyCardDies',
   'card_zone_change | another_card, card_is_creature, card_moves(battlefield, graveyard), card_damaged_by_this': 'thisKillsCreature',
   'life_changed | is_life_gain, affected_player_is(you)': 'youGainLife',
+  'life_changed | is_life_loss, affected_player_is(opp)': 'oppLosesLife',
+  'card_zone_change | controlled_by(you), card_moves(hand, graveyard)': 'youDiscard',
+  'card_zone_change | card_has_subtype(*), card_moves(battlefield, graveyard)': 'cardDiesOfSubtype',
   'spell_cast | another_card, controlled_by(you)': 'youCastSpell',
   'spell_cast | another_card, controlled_by(you), card_has_effect(counter)': 'youCastCounterspell',
 };